Match the new-test NWP fixture to the Int8 ensemble_member contract
main narrowed ensemble_member to Int8 while this branch was open, and the
control-member probe's fixture was written against the old Int64.

+++ packages/ml_core/tests/test_production_helpers.py
@@ -366,7 +366,7 @@ def test_the_guard_rejects_a_hyperparameter_this_code_no_longer_declares(tmp_pat
 
 def _nwp_frame(members: tuple[int, ...]) -> pt.LazyFrame[Nwp]:
     """A minimal NWP frame carrying only the column the probe reads."""
-    frame = pl.LazyFrame({"ensemble_member": list(members)}, schema={"ensemble_member": pl.Int64})
+    frame = pl.LazyFrame({"ensemble_member": list(members)}, schema={"ensemble_member": pl.Int8})
     return cast("pt.LazyFrame[Nwp]", frame)
